Safari-611.1.10/LayoutTests/fast/dom/Range/range-created-during-remove-children-expected.txt

PASS ranges["unload"].startContainer is sample
PASS ranges["unload"].endContainer is sample
PASS ranges["unload"].startOffset is 0
PASS ranges["unload"].endOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].startContainer is sample
PASS ranges["DOMNodeRemovedFromDocument"].endContainer is sample
PASS ranges["DOMNodeRemovedFromDocument"].startOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
